@@ -3,26 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.14.6",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^12.20.15",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.8",
-    "@types/react-router-dom": "^5.1.7",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^13.2.1",
     "firebase": "^8.6.8",
-    "node-sass": "^5.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "node-sass": "^6.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.0",
     "react-icons": "^4.2.0",
     "react-modal": "^3.14.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-toastify": "^7.0.4",
-    "typescript": "^4.3.4",
-    "web-vitals": "^1.1.2"
+    "react-router-dom": "^6.3.0",
+    "react-toastify": "^7.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -49,6 +42,9 @@
     ]
   },
   "devDependencies": {
-    "@types/react-modal": "^3.12.0"
+    "@types/react-modal": "^3.12.0",
+    "typescript": "^4.6.3",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0"
   }
 }
