@@ -1,19 +1,19 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
 
 import './services/firebase';
 import './styles/global.scss';
 
-ReactDOM.render(
-  <React.StrictMode>
+ReactDOM.createRoot(
+  document.getElementById('root')!).render(
+    <React.StrictMode>
     <BrowserRouter>
       <App />
     </BrowserRouter>
   </React.StrictMode>,
-  document.getElementById('root')
-);
+  );
 
 // DOM - Document Object Model
 // ReactDom renderiza o app dentro do HTML
