@@ -1,15 +1,16 @@
-import { Link, useHistory } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { FormEvent, useState } from 'react';
-import illustrationImg from '../assets/images/illustration.svg';
-import logoImg from '../assets/images/logo.svg';
+// import illustrationImg from '../assets/images/illustration.svg';
+// import logoImg from '../assets/images/logo.svg';
 import '../styles/auth.scss';
 import Button from '../components/Button';
 import { database } from '../services/firebase';
 import { useAuth } from '../hooks/useAuth';
+import Logo from '../components/Icon';
 
 function NewRoom() {
   const { user } = useAuth();
-  const history = useHistory();
+  const navigate = useNavigate();
   // Falta recuperar o estado da aplicação p/ conseguir recuperar estado de autenticação, persistência de dados
   const [newRoom, setNewRoom] = useState('');
   async function handleCreateRoom(event: FormEvent) {
@@ -26,20 +27,21 @@ function NewRoom() {
       authorId: user?.id,
     });
 
-    history.push(`/admin/rooms/${firebaseRoom.key}`)
+    navigate(`/admin/rooms/${firebaseRoom.key}`)
 
   }
 
  return (
     <div id="page-auth"> 
       <aside>
-        <img src={illustrationImg} alt="Ilustração simbolizando perguntas e respostas" />
+        {/* <img src={illustrationImg} alt="Ilustração simbolizando perguntas e respostas" /> */}
         <strong>Crie salas de Q&amp;A ao vivo</strong>
         <p>Tire as dúvidas da sua audiência em tempo real</p>
       </aside>
       <main>
         <div className="main-content">
-          <img src={logoImg} alt="Letmeask" />
+          <Logo />
+          {/* <img src={logoImg} alt="Letmeask" /> */}
           <h2>Criar uma nova sala</h2>
           <form onSubmit={ handleCreateRoom }>
             <input
